Initial steps to fix form data-id
