+++ src/controllers/helpers/showTodoBody.js
@@ -16,9 +16,6 @@ export const showTodoBody = (name) => {
     document.getElementsByTagName("table")[0].setAttribute("class", "table table-striped");
     // Iterate thru each todo and create tr for them
     project.todos.forEach(todo => {
-      // Create Add Todo Btn
-      let btn = genrateAddTodoBtn(project);
-      // 
       // Create tr
       let tr = document.createElement("tr");
       let trContent = `
@@ -31,26 +28,23 @@ export const showTodoBody = (name) => {
       `;
       tr.innerHTML = trContent;
       todoBody.appendChild(tr);
-      if (document.getElementById(`addTodoBtn-${project.id}`) === null) {
-        document.getElementById("todosDiv").appendChild(btn);
-      }
     });
   }
 };
 
-const genrateAddTodoBtn = project => {
-  let btn = document.createElement("button");
-  btn.setAttribute("class", "btn btn-sm btn-block btn-primary");
-  btn.setAttribute("id", `addTodoBtn-${project.id}`);
-  btn.setAttribute("data-id", project.id);
-  btn.innerText = "Add New Todo";
-  btn.addEventListener("click", e => {
-    e.stopPropagation();
-    let dataId = e.target.getAttribute("data-id");
-    console.log(dataId);
-    e.target.setAttribute("class", "d-none");
-    document.getElementById("todosSection").setAttribute("class", "mt-3");
-    document.getElementById("todosForm").setAttribute("data-id", project.id);
-  });
-  return btn;
-};
+// const genrateAddTodoBtn = project => {
+//   let btn = document.createElement("button");
+//   btn.setAttribute("class", "btn btn-sm btn-block btn-primary");
+//   btn.setAttribute("id", `addTodoBtn-${project.id}`);
+//   btn.setAttribute("data-id", project.id);
+//   btn.innerText = "Add New Todo";
+//   btn.addEventListener("click", e => {
+//     e.stopPropagation();
+//     let dataId = e.target.getAttribute("data-id");
+//     console.log(dataId);
+//     e.target.setAttribute("class", "d-none");
+//     document.getElementById("todosSection").setAttribute("class", "mt-3");
+//     document.getElementById("todosForm").setAttribute("data-id", project.id);
+//   });
+//   return btn;
+// };
